# Traces: Speed up Activity enumerations by avoiding virtual calls