## Changes

Avoid virtual calls in the IL generated by the allocation-free enumeration helper to speed up enumeration of `Activity` tags, links, & events.

I don't feel like this really needs to be mentioned in the `CHANGELOG` but if anyone feels like it should be there LMK.

## Benchmarks

### Before

|          Method |      Mean |    Error |   StdDev | Allocated |
|---------------- |----------:|---------:|---------:|----------:|
|   EnumerateTags |  30.67 ns | 0.398 ns | 0.458 ns |         - |
|  EnumerateLinks | 227.78 ns | 2.969 ns | 2.777 ns |         - |
| EnumerateEvents | 222.46 ns | 2.740 ns | 2.563 ns |         - |

### After

|          Method |      Mean |    Error |   StdDev | Allocated |
|---------------- |----------:|---------:|---------:|----------:|
|   EnumerateTags |  **28.53 ns** | 0.449 ns | 0.906 ns |         - |
|  EnumerateLinks | **203.03 ns** | 1.731 ns | 1.535 ns |         - |
| EnumerateEvents | **203.18 ns** | 3.395 ns | 3.176 ns |         - |