@CodeBlanch 2022-03-22
Converted this to a draft. I'm talking to runtime about maybe changing the enumerator so we can do this even more efficiently. Depending on that conversation this might see some tweaks.